Factor out SQL setup functions
